requirements.txt
annotated-types==0.7.0
ansitable==0.11.3
anyio==4.7.0
cfgv==3.4.0
click==8.1.7
colored==2.2.4
cycler==0.12.1
distlib==0.3.9
fastapi==0.115.6
filelock==3.16.1
fonttools==4.55.2
h11==0.14.0
identify==2.6.3
idna==3.10
kiwisolver==1.4.7
matplotlib==3.5.1
nodeenv==1.9.1
numpy==1.26.4
packaging==24.2
pgraph-python==0.6.2
pillow==11.0.0
platformdirs==4.3.6
pre_commit==4.0.1
progress==1.6
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
roboticstoolbox-python==1.1.1
rtb-data==1.0.1
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
spatialgeometry==1.1.0
spatialmath-python==1.1.13
starlette==0.41.3
swift-sim==1.1.0
typing_extensions==4.12.2
uvicorn==0.32.1
virtualenv==20.28.0
websockets==14.1
